+++ common/src/main/java/com/evandev/reliable_replacer/api/IProcessedChunk.java
@@ -2,8 +2,14 @@ package com.evandev.reliable_replacer.api;
 
 public interface IProcessedChunk {
     boolean reliableReplacer$hasBeenProcessed();
+
     void reliableReplacer$markProcessed();
 
     boolean reliableReplacer$isDirty();
+
     void reliableReplacer$setDirty(boolean dirty);
+
+    int reliableReplacer$getRulesHash();
+
+    void reliableReplacer$setRulesHash(int hash);
 }